Extract migration processor logic to own abstract class, and move classes around
