+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/action/ApolloV2ToV3MigrationAction.kt
@@ -2,7 +2,7 @@ package com.apollographql.ijplugin.action
 
 import com.apollographql.ijplugin.ApolloBundle
 import com.apollographql.ijplugin.project.apolloProjectService
-import com.apollographql.ijplugin.refactoring.migration.ApolloV2ToV3MigrationProcessor
+import com.apollographql.ijplugin.refactoring.migration.v2tov3.ApolloV2ToV3MigrationProcessor
 import com.apollographql.ijplugin.util.logd
 import com.intellij.openapi.actionSystem.ActionPlaces
 import com.intellij.openapi.actionSystem.ActionUpdateThread
