+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/refactoring/migration/v2tov3/item/AddUseVersion2Compat.kt
@@ -1,5 +1,8 @@
-package com.apollographql.ijplugin.refactoring.migration.item
+package com.apollographql.ijplugin.refactoring.migration.v2tov3.item
 
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItem
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItemUsageInfo
+import com.apollographql.ijplugin.refactoring.migration.item.toMigrationItemUsageInfo
 import com.apollographql.ijplugin.util.addSiblingAfter
 import com.apollographql.ijplugin.util.findPsiFilesByName
 import com.intellij.openapi.project.Project
