+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/refactoring/migration/v2tov3/item/UpdateEnumValueUpperCase.kt
@@ -1,8 +1,10 @@
-package com.apollographql.ijplugin.refactoring.migration.item
+package com.apollographql.ijplugin.refactoring.migration.v2tov3.item
 
 import com.apollographql.ijplugin.refactoring.findClassReferences
 import com.apollographql.ijplugin.refactoring.findFieldReferences
 import com.apollographql.ijplugin.refactoring.findInheritorsOfClass
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItem
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItemUsageInfo
 import com.apollographql.ijplugin.util.unquoted
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiMigration
