+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/refactoring/migration/v2tov3/item/UpdateIdlingResource.kt
@@ -1,6 +1,9 @@
-package com.apollographql.ijplugin.refactoring.migration.item
+package com.apollographql.ijplugin.refactoring.migration.v2tov3.item
 
 import com.apollographql.ijplugin.refactoring.findMethodReferences
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItem
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItemUsageInfo
+import com.apollographql.ijplugin.refactoring.migration.item.toMigrationItemUsageInfo
 import com.apollographql.ijplugin.util.addSiblingBefore
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiMigration
