+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/refactoring/migration/v2tov3/item/UpdateOkHttpExecutionContext.kt
@@ -1,7 +1,10 @@
-package com.apollographql.ijplugin.refactoring.migration.item
+package com.apollographql.ijplugin.refactoring.migration.v2tov3.item
 
 import com.apollographql.ijplugin.refactoring.findClassReferences
 import com.apollographql.ijplugin.refactoring.findOrCreateClass
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItem
+import com.apollographql.ijplugin.refactoring.migration.item.MigrationItemUsageInfo
+import com.apollographql.ijplugin.refactoring.migration.item.toMigrationItemUsageInfo
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiMigration
