+++ intellij-plugin/src/main/resources/messages/ApolloBundle.properties
@@ -27,8 +27,9 @@ action.CompatToOperationBasedCodegenMigrationAction.confirmDialog.message=This w
   Before refactoring please make sure to commit any local changes and that the project builds with no errors.
 
 
+ApolloMigrationRefactoringProcessor.codeReferences=Items to be migrated
+
 ApolloV2ToV3MigrationProcessor.title=Migrate to Apollo Kotlin 3
-ApolloV2ToV3MigrationProcessor.codeReferences=Items to be migrated
 ApolloV2ToV3MigrationProcessor.noUsage=No Apollo Android 2 references found in the project
 
 settings.codegen.automaticCodegenTriggering.text=Automatic code generation
@@ -36,3 +37,7 @@ settings.codegen.automaticCodegenTriggering.comment=Enabling this setting will r
   <code>./gradlew generateApolloSources --continuous</code>.<br>\
   This watches changes to your GraphQL files and re-generates models when they change.
 settings.codegen.title=Code Generation
+
+
+CompatToOperationBasedCodegenMigrationProcessor.title=Migrate to OperationBased Codegen
+CompatToOperationBasedCodegenMigrationProcessor.noUsage=No Compat codegen usage found in the project
